+++ src/entities/comment/model/store.ts
@@ -1,0 +1,5 @@
+import { atom } from "jotai"
+
+export const commentsAtom = atom<any>({});
+export const selectedCommentAtom = atom<any>(null);
+export const newCommentAtom = atom<any>({ body: "", postId: null, userId: 1 });
