+++ src/entities/posts/api/fetchPostsWithTag.ts
@@ -1,0 +1,4 @@
+export const fetchPostsWithTag = async (tag: string) => {
+  const response = await fetch(`/api/posts/tag/${tag}`);
+  return response.json();
+};
