+++ src/entities/search/model/stores.ts
@@ -1,6 +1,15 @@
-import { atomWithLocation } from "jotai-location"
+import { atom } from "jotai"
 
-export const searchParamsAtom = atomWithLocation({
+export interface SearchParamsProps {
+  skip : string;
+  limit : string;
+  search : string;
+  sortBy : string;
+  sortOrder : string;
+  tag : any[];
+}
+
+export const searchParamsAtom : SearchParamsProps = atom({
   skip: "0",
   limit: "10",
   search: '',
