+++ src/features/sort/ui/SelectSortBy.tsx
@@ -1,0 +1,19 @@
+import { useAtom } from "jotai"
+import { sortOrderAtom } from "../model/store.ts"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../../shared/ui"
+
+export default function SelectSortBy() {
+  const [sortOrder, setSortOrder] = useAtom(sortOrderAtom);
+  
+  return (
+    <Select value={sortOrder} onValueChange={setSortOrder}>
+      <SelectTrigger className="w-[180px]">
+        <SelectValue placeholder="정렬 순서" />
+      </SelectTrigger>
+      <SelectContent>
+        <SelectItem value="asc">오름차순</SelectItem>
+        <SelectItem value="desc">내림차순</SelectItem>
+      </SelectContent>
+    </Select>
+  )
+}
