+++ src/shared/hooks/useURLParams.ts
@@ -1,19 +1,24 @@
 import { useAtom } from "jotai"
-import { searchParamsAtom } from "../../entities/search/model/stores.ts"
-import { useCallback } from "react"
+import { searchParamsAtom, SearchParamsProps } from "../../entities/search/model/stores.ts"
 
 export const useURLParams = () => {
   const [params, setParams] = useAtom(searchParamsAtom);
   
-  const updateParams = useCallback((newParams: any) => {
+  const updateParams = (newParams: any) => {
     setParams((prev) => ({
       ...prev,
       ...newParams
     }));
-  }, [setParams]);
+  };
+  
+  // 주소값 초기화
+  const initParams = (newParams : SearchParamsProps) => {
+    setParams(newParams);
+  }
   
   return {
     params,
-    updateParams
+    updateParams,
+    initParams
   };
 };
